+++ CHANGELOG.md
@@ -103,6 +103,7 @@
 
 - Wlogout theme - `hadi493` adapted from `hadi493/wlogout` repo
   - Fully acreditied in source files
+- Wlogout theme from `hadi493` but icons from `LordWorm1996`
 - `Hyprland - OEM Default animation`
   - A simple low overhead animation from the default LUA file
 - `wlogout` themese